Added report issue form to main navigation drawer
