+++ contentcuration/contentcuration/frontend/shared/views/MainNavigationDrawer.vue
@@ -1,70 +1,83 @@
 <template>
 
-  <VNavigationDrawer
-    v-model="drawer"
-    fixed
-    temporary
-    style="z-index: 1000;"
-  >
-    <VToolbar color="primary" dark>
-      <VBtn flat icon @click="drawer = false">
-        <Icon>clear</Icon>
-      </VBtn>
-      <VToolbarTitle class="notranslate">
-        Kolibri Studio Beta
-      </VToolbarTitle>
-    </VToolbar>
-    <VList>
-      <VListTile :href="channelsLink">
-        <VListTileAction>
-          <Icon>home</Icon>
-        </VListTileAction>
-        <VListTileContent class="subheading">
-          <VListTileTitle>{{ $tr('channelsLink') }}</VListTileTitle>
-        </VListTileContent>
-      </VListTile>
-      <VListTile v-if="user.is_admin" :href="administrationLink">
-        <VListTileAction>
-          <Icon>people</Icon>
-        </VListTileAction>
-        <VListTileContent class="subheading">
-          <VListTileTitle>{{ $tr('administrationLink') }}</VListTileTitle>
-        </VListTileContent>
-      </VListTile>
-      <VListTile :href="settingsLink">
-        <VListTileAction>
-          <Icon>settings</Icon>
-        </VListTileAction>
-        <VListTileContent class="subheading">
-          <VListTileTitle>{{ $tr('settingsLink') }}</VListTileTitle>
-        </VListTileContent>
-      </VListTile>
-      <VListTile :href="helpLink" target="_blank">
-        <VListTileAction>
-          <Icon>open_in_new</Icon>
-        </VListTileAction>
-        <VListTileContent class="subheading">
-          <VListTileTitle>{{ $tr('helpLink') }}</VListTileTitle>
-        </VListTileContent>
-      </VListTile>
-      <VListTile @click="logout">
-        <VListTileAction>
-          <Icon>exit_to_app</Icon>
-        </VListTileAction>
-        <VListTileContent class="subheading">
-          <VListTileTitle>{{ $tr('logoutLink') }}</VListTileTitle>
-        </VListTileContent>
-      </VListTile>
-    </VList>
-    <VContainer>
-      <KolibriLogo :height="75" />
-      <ActionLink
-        :text="$tr('copyright', {year: new Date().getFullYear()})"
-        href="https://learningequality.org/"
-        target="_blank"
-      />
-    </VContainer>
-  </VNavigationDrawer>
+  <div>
+    <VNavigationDrawer
+      v-model="drawer"
+      fixed
+      temporary
+      style="z-index: 1000;"
+    >
+      <VToolbar color="primary" dark>
+        <VBtn flat icon @click="drawer = false">
+          <Icon>clear</Icon>
+        </VBtn>
+        <VToolbarTitle class="notranslate">
+          Kolibri Studio Beta
+        </VToolbarTitle>
+      </VToolbar>
+      <VList>
+        <VListTile :href="channelsLink">
+          <VListTileAction>
+            <Icon>home</Icon>
+          </VListTileAction>
+          <VListTileContent class="subheading">
+            <VListTileTitle>{{ $tr('channelsLink') }}</VListTileTitle>
+          </VListTileContent>
+        </VListTile>
+        <VListTile v-if="user.is_admin" :href="administrationLink">
+          <VListTileAction>
+            <Icon>people</Icon>
+          </VListTileAction>
+          <VListTileContent class="subheading">
+            <VListTileTitle>{{ $tr('administrationLink') }}</VListTileTitle>
+          </VListTileContent>
+        </VListTile>
+        <VListTile :href="settingsLink">
+          <VListTileAction>
+            <Icon>settings</Icon>
+          </VListTileAction>
+          <VListTileContent class="subheading">
+            <VListTileTitle>{{ $tr('settingsLink') }}</VListTileTitle>
+          </VListTileContent>
+        </VListTile>
+        <VListTile :href="helpLink" target="_blank">
+          <VListTileAction>
+            <Icon>open_in_new</Icon>
+          </VListTileAction>
+          <VListTileContent class="subheading">
+            <VListTileTitle>{{ $tr('helpLink') }}</VListTileTitle>
+          </VListTileContent>
+        </VListTile>
+        <VListTile @click="logout">
+          <VListTileAction>
+            <Icon>exit_to_app</Icon>
+          </VListTileAction>
+          <VListTileContent class="subheading">
+            <VListTileTitle>{{ $tr('logoutLink') }}</VListTileTitle>
+          </VListTileContent>
+        </VListTile>
+      </VList>
+      <VContainer>
+        <KolibriLogo :height="75" />
+        <ActionLink
+          :text="$tr('copyright', {year: new Date().getFullYear()})"
+          href="https://learningequality.org/"
+          target="_blank"
+        />
+        <p class="mt-4">
+          <ActionLink
+            :text="$tr('reportIssue')"
+            @click="showReportForm"
+          />
+        </p>
+      </VContainer>
+    </VNavigationDrawer>
+    <!-- Overwrite dark mode -->
+    <span style="color: var(--v-grey-darken4);">
+      <ReportIssueForm v-model="showIssueReportForm" />
+    </span>
+
+  </div>
 
 </template>
 
@@ -73,17 +86,24 @@
 
   import { mapActions, mapState } from 'vuex';
   import KolibriLogo from './KolibriLogo';
+  import ReportIssueForm from 'shared/views/errors/ReportIssueForm';
 
   export default {
     name: 'MainNavigationDrawer',
     components: {
       KolibriLogo,
+      ReportIssueForm,
     },
     props: {
       value: {
         type: Boolean,
         default: false,
       },
+    },
+    data() {
+      return {
+        showIssueReportForm: false,
+      };
     },
     computed: {
       ...mapState({
@@ -112,6 +132,10 @@
     },
     methods: {
       ...mapActions(['logout']),
+      showReportForm() {
+        this.showIssueReportForm = true;
+        this.drawer = false;
+      },
     },
     $trs: {
       channelsLink: 'Channels',
@@ -120,6 +144,7 @@
       helpLink: 'Help and support',
       logoutLink: 'Sign out',
       copyright: '© {year} Learning Equality',
+      reportIssue: 'Report an issue',
     },
   };
